+++ History.markdown
@@ -2,6 +2,7 @@
 
 ### Bug Fixes
 
+  * Fix wildcard `exclude`/`include` patterns not matching files several directories below the matched directory, e.g. in collections (#10011)
   * Avoid caching resource when called via `include_relative` tag (#9784)
   * Fix logs containing IPv6 URLs (#9813)
   * Do not treat colons in `url_placeholders` as URI delimiters (#9850)